% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bcsd_obs}
\alias{bcsd_obs}
\title{Monthly Gridded Meteorological Observations}
\format{
An object of class \code{stars_proxy} (inherits from \code{stars}) of dimension 81 x 33 x 12.
}
\usage{
bcsd_obs
}
\description{
These are the monthly observational data used for BCSD downscaling. See: <https://gdo-dcp.ucllnl.org/downscaled_cmip_projections/dcpInterface.html#About> for more information." ;
"Atmospheric Temperature, Air Temperature Atmosphere, Precipitation, Rain, Maximum Daily Temperature, Minimum  Daily Temperature" ;
}
\keyword{datasets}
