% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{as}
\alias{as}
\alias{coerce,stars,Raster-method}
\alias{coerce,stars_proxy,Raster-method}
\alias{coerce,stars,Terra-method}
\alias{coerce,stars_proxy,Terra-method}
\title{Coerce stars object into a RasterLayer, RasterBrick, or SpatRaster}
\usage{
\S4method{coerce}{stars,Raster}(from, to = "Raster", strict = TRUE)

\S4method{coerce}{stars,SpatRaster}(from, to = "SpatRaster", strict = TRUE)
}
\value{
RasterLayer or RasterBrick

SpatRaster
}
\description{
Coerce stars object into a RasterLayer, RasterBrick, or SpatRaster
}
\details{
If the stars object has more than three dimensions, all dimensions higher than the third will be collapsed into the third dimensions. If the stars object has only an x/y raster but multiple attributes, these are merged first, then put in a raster brick or SpatRaster.
}
