% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{cut_stars}
\alias{cut_stars}
\alias{cut.array}
\alias{cut.matrix}
\alias{cut.stars}
\title{cut methods for stars objects}
\usage{
\method{cut}{array}(x, breaks, ...)

\method{cut}{matrix}(x, breaks, ...)

\method{cut}{stars}(x, breaks, ...)
}
\arguments{
\item{x}{see \link[base]{cut}}

\item{breaks}{see \link[base]{cut}}

\item{...}{see \link[base]{cut}}
}
\description{
cut methods for stars objects
}
\details{
R's \code{factor} only works for vectors, not for arrays or matrices. This is a work-around (or hack?) to keep the factor levels generated by \code{cut} and use them in plots.
}
