% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{dplyr}
\alias{dplyr}
\alias{filter.stars}
\alias{dplyr}
\alias{mutate.stars}
\alias{dplyr}
\alias{select.stars}
\alias{dplyr}
\alias{pull.stars}
\alias{dplyr}
\alias{as.tbl_cube.stars}
\title{dplyr verbs for stars objects}
\usage{
filter.stars(.data, ...)

mutate.stars(.data, ...)

select.stars(.data, ...)

pull.stars(.data, var = -1)

as.tbl_cube.stars(x, ...)
}
\arguments{
\item{.data}{see \link[dplyr]{filter}}

\item{...}{see \link[dplyr]{filter}}

\item{var}{see \link[dplyr]{pull}}

\item{x}{object of class \code{stars}}
}
\description{
dplyr verbs for stars objects
}
