% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{expand_dimensions}
\alias{expand_dimensions}
\title{expand the dimension values into a list}
\usage{
expand_dimensions(x, ...)
}
\arguments{
\item{x}{object of class `stars` or `dimensions`}

\item{...}{ignored}
}
\description{
expand the dimension values into a list
}
