% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blocks.R
\name{get_blocks}
\alias{get_blocks}
\title{Specify parameters to load raster in blocks}
\usage{
get_blocks(img_rows, img_cols, x_window, y_window)
}
\arguments{
\item{img_rows}{number of input raster rows (integer)}

\item{img_cols}{number of input raster columns (integer)}

\item{x_window}{number of rows in block (integer)}

\item{y_window}{number of columns in block (integer)}
}
\value{
matrix with specified \code{nXOff}, \code{nYOff}, \code{nXsize},
and \code{nYSize} parameters for every block
}
\description{
Helper function for specifying the block parameters (\code{nXOff},
\code{nYOff}, \code{nXsize}, and \code{nYSize}) required by \code{RasterIO}
argument in \link{read_stars}
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
r = read_stars(tif, proxy = TRUE)
blocks = get_blocks(nrow(r), ncol(r), 256L, 256L)
for (i in seq_len(nrow(blocks))) {
  io = list(nXOff = blocks[i, 1], nYOff = blocks[i, 2],
            nXSize = blocks[i, 3], nYSize = blocks[i, 4])
  x = read_stars(tif, proxy = FALSE, RasterIO = io)
}

}
