% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{\%in\%,stars-method}
\alias{\%in\%,stars-method}
\title{evaluate whether cube values are in a given set}
\usage{
\S4method{\%in\%}{stars}(x, table)
}
\description{
evaluate whether cube values are in a given set
}
