% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{L7_ETMs}
\alias{L7_ETMs}
\title{Landsat-7 bands for a selected region around Olinda, BR}
\format{
An object of class \code{stars_proxy} (inherits from \code{stars}) of dimension 349 x 352 x 6.
}
\usage{
L7_ETMs
}
\description{
Probably containing the six 30 m bands:
\itemize{
 \item Band 1 Visible (0.45 - 0.52 µm) 30 m
 \item Band 2 Visible (0.52 - 0.60 µm) 30 m
 \item Band 3 Visible (0.63 - 0.69 µm) 30 m
 \item Band 4 Near-Infrared (0.77 - 0.90 µm) 30 m
 \item Band 5 Short-wave Infrared (1.55 - 1.75 µm) 30 m
 \item Band 7 Mid-Infrared (2.08 - 2.35 µm) 30 m
}
}
\keyword{datasets}
