% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{make_intervals}
\alias{make_intervals}
\title{create an intervals object}
\usage{
make_intervals(start, end)
}
\arguments{
\item{start}{vector with start values, or 2-column matrix with start and end values in column 1 and 2, respectively}

\item{end}{vector with end values}
}
\description{
create an intervals object, assuming left-closed and right-open intervals
}
