% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Math.stars}
\alias{Math.stars}
\title{Mathematical operations for stars objects}
\usage{
\method{Math}{stars}(x, ...)
}
\arguments{
\item{x}{object of class stars}

\item{...}{parameters passed on to the Math functions}
}
\description{
Mathematical operations for stars objects
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = st_stars(tif)
a = sqrt(x)
b = log(x, base = 10)
}
