% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdim.R
\name{mdim}
\alias{mdim}
\alias{read_mdim}
\alias{write_mdim}
\title{Read or write data using GDAL's multidimensional array API}
\usage{
read_mdim(
  filename,
  variable = character(0),
  ...,
  options = character(0),
  raster = NULL,
  offset = integer(0),
  count = integer(0),
  step = integer(0),
  proxy = FALSE,
  debug = FALSE,
  bounds = TRUE
)

write_mdim(
  x,
  filename,
  driver = detect.driver(filename),
  ...,
  root_group_options = character(0),
  options = character(0),
  as_float = TRUE
)
}
\arguments{
\item{filename}{name of the source or destination file or data source}

\item{variable}{name of the array to be read}

\item{...}{ignored}

\item{options}{character; driver specific options regarding the opening (read_mdim) or creation (write_mdim) of the dataset}

\item{raster}{names of the raster variables (default: first two dimensions)}

\item{offset}{integer; offset for each dimension (pixels) of sub-array to read (default: 0,0,0,...) (requires sf >= 1.0-9)}

\item{count}{integer; size for each dimension (pixels) of sub-array to read (default: read all) (requires sf >= 1.0-9)}

\item{step}{integer; step size for each dimension (pixels) of sub-aray to read (requires sf >= 1.0-9)}

\item{proxy}{logical; return proxy object? (not functional yet)}

\item{debug}{logical; print debug info?}

\item{bounds}{logical or character: if \code{TRUE} tries to infer from "bounds" attribute; if character, 
named vector of the form \code{c(longitude="lon_bnds", latitude="lat_bnds")} with names dimension names}

\item{x}{stars object}

\item{driver}{character; driver name}

\item{root_group_options}{character; driver specific options regarding the creation of the root group}

\item{as_float}{logical; if \code{TRUE} write 4-byte floating point numbers, if \code{FALSE} write 8-byte doubles}
}
\description{
Read or write data using GDAL's multidimensional array API
}
\details{
it is assumed that the first two dimensions are easting / northing
}
\examples{
set.seed(135)
m = matrix(runif(10), 2, 5)
names(dim(m)) = c("stations", "time")
times = as.Date("2022-05-01") + 1:5
pts = st_as_sfc(c("POINT(0 1)", "POINT(3 5)"))
s = st_as_stars(list(Precipitation = m)) |>
 st_set_dimensions(1, values = pts) |>
 st_set_dimensions(2, values = times)
nc = tempfile(fileext=".nc")
if (compareVersion(sf_extSoftVersion()["GDAL"], "3.4.0") > -1) {
  write_mdim(s, nc)
  # try ncdump on the generated file
  print(read_mdim(nc))
}
}
