% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{merge}
\alias{merge}
\alias{split.stars}
\alias{split}
\alias{merge.stars}
\title{merge or split stars object}
\usage{
\method{split}{stars}(x, f = length(dim(x)), drop = TRUE, ...)

\method{merge}{stars}(x, y, ..., name = "attributes")
}
\arguments{
\item{x}{object of class \code{stars}}

\item{f}{the name or index of the dimension to split; by default the last dimension}

\item{drop}{ignored}

\item{...}{if defined, the first unnamed argument is used for dimension values, if not defined, attribute names are used for dimension values}

\item{y}{needs to be missing}

\item{name}{name for the new dimension}
}
\value{
merge merges attributes of a stars object into a new dimension; split splits a dimension over attributes
}
\description{
merge attributes into a dimension, or split a dimension over attributes
}
\details{
split.stars works on the first attribute, and will give an error when more than one attribute is present
}
