% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Ops.stars}
\alias{Ops.stars}
\title{S3 Ops Group Generic Functions for stars objects}
\usage{
\method{Ops}{stars}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{stars}}

\item{e2}{object of class \code{stars}}
}
\value{
object of class \code{stars}
}
\description{
Ops functions for stars objects, including comparison, product and divide, add, subtract
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = st_stars(tif)
x * x
x / x
x + x
x + 10
}
