% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stars}
\alias{plot.stars}
\alias{plot.stars}
\alias{image.stars}
\title{plot stars object, with subplots for each level of first non-spatial dimension}
\usage{
\method{plot}{stars}(x, y, ..., join_zlim = TRUE, main = names(x)[1])

\method{image}{stars}(x, ..., band = 1, attr = 1, asp = 1, rgb = NULL,
  maxColorValue = 1, xlab = names(dims)[1], ylab = names(dims)[2],
  xlim = st_bbox(x)$xlim, ylim = st_bbox(x)$ylim, useRaster = TRUE,
  text_values = FALSE)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{y}{ignored}

\item{...}{passed on to \code{image.default}}

\item{join_zlim}{logical; if \code{TRUE}, compute a single zlim for all subplots from array range}

\item{main}{character; subplot title prefix; use \code{""} to get only time, use \code{NULL} to suppress subplot titles}

\item{band}{integer; which band (dimension) to plot}

\item{attr}{integer; which attribute to plot}

\item{asp}{numeric; aspect ratio of image}

\item{rgb}{integer; specify three bands to form an rgb composite}

\item{maxColorValue}{numeric; passed on to \link{rgb}}

\item{xlab}{character; x axis label}

\item{ylab}{character; y axis label}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}

\item{useRaster}{logical; see \link{image.default}}

\item{text_values}{logical; print values as text on image?}
}
\description{
plot stars object, with subplots for each level of first non-spatial dimension
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = st_stars(tif)
image(x, col = grey((3:9)/10))
}
