% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R, R/proxy.R
\name{predict.stars}
\alias{predict.stars}
\alias{predict.stars_proxy}
\title{Predict values, given a model object, for a stars or stars_proxy object}
\usage{
\method{predict}{stars}(object, model, ...)

\method{predict}{stars_proxy}(object, model, ...)
}
\arguments{
\item{object}{object of class `stars`}

\item{model}{model object of a class that has a predict method; check with `methods(class = class(object))`}

\item{...}{arguments passed on to this predict method}
}
\description{
Predict values, given a model object, for a stars or stars_proxy object
}
\details{
separate predictors in object need to be separate attributes in object; 
in case they are e.g. in a band dimension, use `split(object)`
}
