% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R, R/stars.R
\name{print_stars}
\alias{print_stars}
\alias{as.data.frame.dimensions}
\alias{print.dimensions}
\alias{print.stars}
\title{print stars or dimensions object}
\usage{
\method{as.data.frame}{dimensions}(
  x,
  ...,
  digits = max(3, getOption("digits") - 3),
  usetz = TRUE,
  stars_crs = getOption("stars.crs") \%||\% 28,
  all = FALSE
)

\method{print}{dimensions}(x, ...)

\method{print}{stars}(x, ..., n = 1e+05, abbrev = 30)
}
\arguments{
\item{x}{object of class stars or of class dimensions}

\item{...}{passed on to \code{as.data.frame.dimensions}}

\item{digits}{number of digits to print numbers}

\item{usetz}{logical; used to format \code{POSIXct} values}

\item{stars_crs}{maximum width of string for CRS objects}

\item{all}{logical; if \code{TRUE} print also fields entirely filled with \code{NA} or \code{NULL}}

\item{n}{when prod(dim(x)) > 10 * n, the first n cells are used for attribute summary statistics}

\item{abbrev}{number of characters to abbreviate attribute names to}
}
\description{
print stars or dimensions object
}
