% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mdim}
\alias{read_mdim}
\title{Read data using GDAL's multidimensional array API (experimental)}
\usage{
read_mdim(
  x,
  variable = character(0),
  ...,
  options = character(0),
  raster = NULL,
  offset = integer(0),
  count = integer(0),
  step = integer(0),
  proxy = FALSE,
  debug = FALSE
)
}
\arguments{
\item{x}{data source name}

\item{variable}{name of the array to be read}

\item{...}{ignored}

\item{options}{array opening options}

\item{raster}{names of the raster variables (default: first two)}

\item{offset}{integer; offset for each dimension (pixels) of sub-array to read (default: 0,0,0,...) (requires sf >= 1.0-9)}

\item{count}{integer; size for each dimension (pixels) of sub-array to read (default: read all) (requires sf >= 1.0-9)}

\item{step}{integer; step size for each dimension (pixels) of sub-aray to read (requires sf >= 1.0-9)}

\item{proxy}{logical; return proxy object? (not functional yet)}

\item{debug}{logical; print debug info?}
}
\description{
Read data using GDAL's multidimensional array API (experimental)
}
\details{
it is assumed that the first two dimensions are easting / northing
}
