% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{st_transform_proj}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lwgeom}{\code{\link[lwgeom]{st_transform_proj}}}
}}

