% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{st_apply}
\alias{st_apply}
\alias{st_apply.stars}
\title{st_apply apply a function to one or more array dimensions}
\usage{
\method{st_apply}{stars}(X, MARGIN, FUN, ...,
  CLUSTER = parallel::getDefaultCluster())
}
\arguments{
\item{X}{object of class \code{stars}}

\item{MARGIN}{see \link[base]{apply}; if \code{MARGIN} is a character vector,}

\item{FUN}{see \link[base]{apply}}

\item{...}{arguments passed on to \code{FUN}}

\item{CLUSTER}{cluster to use for parallel apply; see \link[parallel]{makeCluster}}
}
\value{
object of class \code{stars} with accordingly reduced number of dimensions; in case \code{FUN} returns more than one value, a new dimension is created carrying the name of the function used; see the examples.
}
\description{
st_apply apply a function to array dimensions: aggregate over space, time, or something else
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = read_stars(tif)
st_apply(x, 1:2, mean) # mean band value for each pixel
st_apply(x, 3, mean)   # mean of all pixels for each band
st_apply(x, 1:2, range) # min and max band value for each pixel
}
