% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{st_as_stars.SpatRaster}
\alias{st_as_stars.SpatRaster}
\title{#' @export
st_as_stars.SpatRaster = function(.x, ...) {
	if (!requireNamespace("terra", quietly = TRUE))
		stop("package terra required, please install it first") # nocov
	if (!requireNamespace("raster", quietly = TRUE))
		stop("package terra required, please install it first") # nocov
	st_as_stars(as(.x, "Raster"), ...)
}}
\usage{
\method{st_as_stars}{SpatRaster}(.x, ..., att = 1, ignore_file = FALSE)
}
\description{
#' @export
st_as_stars.SpatRaster = function(.x, ...) {
	if (!requireNamespace("terra", quietly = TRUE))
		stop("package terra required, please install it first") # nocov
	if (!requireNamespace("raster", quietly = TRUE))
		stop("package terra required, please install it first") # nocov
	st_as_stars(as(.x, "Raster"), ...)
}
}
