% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_cells}
\alias{st_cells}
\title{return the cell index corresponding to the location of a set of points}
\usage{
st_cells(x, sf)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{sf}{object of class \code{sf} or \code{sfc}}
}
\description{
return the cell index corresponding to the location of a set of points
}
