% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_coordinates}
\alias{st_coordinates}
\alias{st_coordinates.stars}
\alias{as.data.frame.stars}
\title{retrieve coordinates for raster or vector cube cells}
\usage{
\method{st_coordinates}{stars}(x, ..., add_max = FALSE)

\method{as.data.frame}{stars}(x, ..., add_max = FALSE)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{...}{ignored}

\item{add_max}{logical; if \code{TRUE}, dimensions are given with a min (x) and max (x_max) value}
}
\description{
retrieve coordinates for raster or vector cube cells
}
