% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R, R/proxy.R
\name{st_crop}
\alias{st_crop}
\alias{st_crop.stars}
\alias{st_crop.stars_proxy}
\title{crop a stars object}
\usage{
\method{st_crop}{stars}(x, y, ..., crop = TRUE)

\method{st_crop}{stars_proxy}(x, y, ..., crop = TRUE)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{y}{object of class \code{sf}, \code{sfc} or \code{bbox}}

\item{...}{ignored}

\item{crop}{logical; if \code{TRUE}, the spatial extent of the returned object is cropped to still cover \code{obj}}
}
\description{
crop a stars object
}
