% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{st_cube_view_to_dimensions}
\alias{st_cube_view_to_dimensions}
\title{convert cube_view object to stars dimensions object}
\usage{
st_cube_view_to_dimensions(v, ...)
}
\arguments{
\item{v}{object of class \code{cube_view} (package gdalcubes)}

\item{...}{ignored}
}
\description{
convert cube_view object to stars dimensions object
}
\details{
wronly interprets time, at the moment, and only for P1M (monthly)
}
\examples{

library(gdalcubes)
example(cube_view)
st_cube_view_to_dimensions(vnew)
st_get_dimension_values(st_cube_view_to_dimensions(vnew), 1)
st_get_dimension_values(st_cube_view_to_dimensions(vnew), 2)
st_get_dimension_values(st_cube_view_to_dimensions(vnew), 3)

}
