% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{st_dimensions}
\alias{st_dimensions}
\alias{st_dimensions.stars}
\alias{st_dimensions.array}
\alias{st_dimensions.default}
\alias{st_set_dimensions}
\title{get dimensions from stars object}
\usage{
st_dimensions(.x, ...)

\method{st_dimensions}{stars}(.x, ...)

\method{st_dimensions}{array}(.x, ...)

\method{st_dimensions}{default}(.x, ..., .raster = rep(NA_character_, 2),
  cell_midpoints = FALSE)

st_set_dimensions(.x, which, values, names)
}
\arguments{
\item{.x}{object to retrieve dimensions information from}

\item{...}{further arguments}

\item{.raster}{length 2 character array with names (if any) of the raster dimensions}

\item{cell_midpoints}{logical; if \code{TRUE} AND the dimension values are strictly regular, the values are interpreted as the cell midpoint values rather than the cell offset values when calculating offset (i.e., the half-cell-size correction is applied); can have a value for each dimension, or else is recycled}

\item{which}{integer which dimension to change}

\item{values}{values for this dimension (e.g. \code{sfc} list-column)}

\item{names}{character; new names vector for (all) dimensions, ignoring \code{which}}
}
\value{
the \code{dimensions} attribute of \code{x}, of class \code{dimensions}
}
\description{
get dimensions from stars object
}
