% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{st_geotransform}
\alias{st_geotransform}
\alias{st_geotransform<-}
\alias{st_geotransform<-.stars}
\title{get or set the geotransform, or rotation matrix}
\usage{
st_geotransform(x, ...)

st_geotransform(x) <- value

\method{st_geotransform}{stars}(x) <- value
}
\arguments{
\item{x}{object of class stars or dimensions}

\item{...}{ignored}

\item{value}{length 6 numeric vector, or 2 x 2 (scaled) rotation matrix}
}
\description{
get or set the geotransform, or rotation matrix
}
\examples{
l = st_as_stars(L7_ETMs)
rot = function(theta) { 
   th = theta / 180 * pi
   matrix(c(cos(th), sin(th), -sin(th), cos(th)), 2) 
}
st_geotransform(l) = rot(20) * 28.5 # clockwise, 20 degrees, scale by cell size
if (interactive()) {
  plot(l[,,,1])
}
}
