% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{st_join.stars}
\alias{st_join.stars}
\title{Spatially join a stars and an `sf` object}
\usage{
st_join.stars(x, y, ..., join = st_intersects, what = "left",
  as_points = NA, warn = TRUE)
}
\arguments{
\item{x}{object of class stars}

\item{y}{object of class sf, or one that can be coerced into that by \link{st_as_sf}}

\item{...}{arguments that will be passed on to the join function}

\item{join}{join function, which should return an sgbp object}

\item{what}{either "left1" or "right"}

\item{as_points}{logical; controls whether grid cells in \code{x} will be treated as points, or as cell areas; \link{st_intersects.stars} by default will derive this from \code{x}'s metadata, or else assume areas.}

\item{warn}{logical; if TRUE, warn on 1-to-many matches when \code{what} is \code{"left1"}}
}
\description{
Spatially join a stars and an `sf` object
}
