% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R, R/subset.R
\name{st_normalize}
\alias{st_normalize}
\alias{st_normalize.stars_proxy}
\alias{st_normalize.stars}
\title{Move raster offset such that grid start at row/col (1,1)}
\usage{
\method{st_normalize}{stars_proxy}(x, domain = c(0, 0, 1, 1), ...)

\method{st_normalize}{stars}(x, domain = c(0, 0, 1, 1), ...)
}
\arguments{
\item{x}{object of class `stars` or `stars_proxy`}

\item{domain}{ignored}

\item{...}{ignored}
}
\description{
Move raster offset such that grid start at row/col (1,1)
}
\details{
the `stars_proxy` method does nothing, but raises an error when a raster does not start at cell (1,1)

the `stars` method "moves" the `offset` of the raster dimensions, such that the `from`
field becomes `(1,1)` for the raster dimensions; as a side effect, it moves the raster
dimensions `x` and `y` to first and second position.  If the `from` field is already `(1,1)` or the
object does not have raster dimensions it does nothing.
}
