% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{st_read_crs}
\alias{st_read_crs}
\title{read coordinate reference system from GDAL data set}
\usage{
st_read_crs(file, options = character(0))
}
\arguments{
\item{file}{character; file name}

\item{options}{character; raster layer read options}
}
\value{
object of class \code{crs}, see \link[sf]{st_crs}.
}
\description{
read coordinate reference system from GDAL data set
}
