% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp.R
\name{st_rotate}
\alias{st_rotate}
\title{Transform rotated long/lat to regular long/lat}
\usage{
st_rotate(.x, lon0, lat0, north = TRUE)
}
\arguments{
\item{.x}{object of class \code{stars}}

\item{lon0}{longitude of the rotated pole}

\item{lat0}{latitude of the rotated pole}

\item{north}{logical; if \code{TRUE} the pole refers to the North pole, otherwise the South pole}
}
\value{
curvilinear stars object with coordinates in regular long/lat (North pole at lat=90)
}
\description{
Transform rotated long/lat to regular long/lat
}
