% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_set_bbox}
\alias{st_set_bbox}
\title{set bounding box parameters of regular grid}
\usage{
st_set_bbox(x, value, ...)
}
\arguments{
\item{x}{object of class dimensions, stars or stars_proxy}

\item{value}{object of class bbox}

\item{...}{ignored}
}
\description{
set bounding box parameters of regular grid
}
