% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.R
\name{st_sfc2xy}
\alias{st_sfc2xy}
\title{replace POINT simple feature geometry list with an x y raster}
\usage{
st_sfc2xy(x, ...)
}
\arguments{
\item{x}{object of class \code{stars}, or of class \code{sf}}

\item{...}{passed on to \link{as.data.frame.stars}}
}
\value{
object of class \code{stars} with a POINT list replaced by x and y raster dimensions. This only works when the points are distributed over a regular or rectilinear grid.
}
\description{
replace POINT simple feature geometry list with an x y raster
}
