% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{st_write.stars}
\alias{st_write.stars}
\title{write stars object to gdal dataset (typically: to file)}
\usage{
\method{st_write}{stars}(obj, dsn, layer = 1, ...,
  driver = detect.driver(dsn), options = character(0), type = "Float32",
  NA_value = NA_real_)
}
\arguments{
\item{obj}{object of class \code{stars}}

\item{dsn}{gdal dataset (file) name}

\item{layer}{attribute name; if missing, the first attribute is written}

\item{...}{passed on to \link[sf]{gdal_write}}

\item{driver}{driver driver name; see \link[sf]{st_drivers}}

\item{options}{character vector with options}

\item{type}{character; output binary type, one of: \code{Byte} for eight bit unsigned integer, \code{UInt16} for sixteen bit unsigned integer, \code{Int16} for sixteen bit signed integer, \code{UInt32} for thirty two bit unsigned integer, \code{Int32} for thirty two bit signed integer, \code{Float32} for thirty two bit floating point, \code{Float64} for sixty four bit floating point.}

\item{NA_value}{non-NA value that should represent R's \code{NA} value in the target raster file; if set to \code{NA}, it will be ignored.}
}
\description{
write stars object to gdal dataset (typically: to file)
}
