% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars-package.R
\docType{package}
\name{stars-package}
\alias{stars-package}
\alias{_PACKAGE}
\alias{stars}
\title{stars: Spatiotemporal Arrays, Raster and Vector Data Cubes}
\description{
Package `stars` provides infrastructure for _data cubes_, array
data with labeled dimensions, with emphasis on arrays where some
of the dimensions relate to time and/or space.
Learn more in the \href{https://r-spatial.github.io/stars/}{introductory vignettes}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatial.github.io/stars/}
  \item \url{https://github.com/r-spatial/stars/}
  \item Report bugs at \url{https://github.com/r-spatial/stars/issues/}
}

}
\author{
\strong{Maintainer}: Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})

Other contributors:
\itemize{
  \item Michael Sumner (\href{https://orcid.org/0000-0002-2471-7511}{ORCID}) [contributor]
  \item Etienne Racine [contributor]
  \item Adriano Fantini [contributor]
  \item David Blodgett [contributor]
}

}
