% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{stars_extSoftVersion}
\alias{stars_extSoftVersion}
\title{Provide the external dependencies versions of the libraries linked to stars}
\usage{
stars_extSoftVersion()
}
\description{
Provide the external dependencies versions of the libraries linked to stars
}
