% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stars_sentinel2}
\alias{stars_sentinel2}
\title{Sentinel-2 sample tile}
\format{
An object of class \code{stars_proxy} (inherits from \code{stars}) of dimension 10980 x 10980 x 4.
}
\usage{
stars_sentinel2
}
\description{
Sentinel-2 sample tile, downloaded from <https://scihub.copernicus.eu/>
reads the four 10-m bands: B2 (490 nm), B3 (560 nm), B4 (665 nm) and B8 (842 nm)
}
\keyword{datasets}
