% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdim.R
\name{write_mdim}
\alias{write_mdim}
\title{Write stars object using GDAL multidimensional array interface}
\usage{
write_mdim(x, filename, driver = detect.driver(filename), ...)
}
\arguments{
\item{x}{stars object}

\item{filename}{destination file name}

\item{driver}{character; driver name}

\item{...}{ignored}
}
\description{
Write stars object using GDAL multidimensional array interface
}
