% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.stars}
\alias{aggregate.stars}
\title{spatially aggregate stars object to feature geometry}
\usage{
\method{aggregate}{stars}(x, by, FUN, ..., drop = FALSE,
  dim = which_sfc(x)[1], join = st_intersects)
}
\arguments{
\item{x}{object of class \code{stars} with information to be aggregated}

\item{by}{object of class \code{stars} with aggregation geometry}

\item{FUN}{aggregation function}

\item{...}{arguments passed on to \code{FUN}}

\item{drop}{logical; ignored}

\item{dim}{integer; geometry dimension that will be aggregated}

\item{join}{join function to find matches of x to by}
}
\description{
spatially aggregate stars object to feature geometry
}
