% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{as}
\alias{as}
\alias{coerce,stars,Raster-method}
\title{Coerce stars object into a Raster raster or brick}
\arguments{
\item{from}{object to coerce}
}
\description{
Coerce stars object into a Raster raster or brick
}
