% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{as}
\alias{as}
\alias{coerce,stars,Raster-method}
\alias{coerce,stars_proxy,Raster-method}
\title{Coerce stars object into a Raster raster or brick}
\arguments{
\item{from}{object to coerce}
}
\value{
RasterLayer or RasterBrick
}
\description{
Coerce stars object into a Raster raster or brick
}
\details{
If the stars object has more than three dimensions, all dimensions higher than the third will be collapsed into the third dimensions. If the stars object has only an x/y raster but multiple attributes, these are merged first, then put in a raster brick.
}
