% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{dplyr}
\alias{dplyr}
\alias{filter.stars}
\alias{dplyr}
\alias{mutate.stars}
\alias{dplyr}
\alias{select.stars}
\alias{dplyr}
\alias{pull.stars}
\alias{dplyr}
\alias{as.tbl_cube.stars}
\title{dplyr verbs for stars objects}
\usage{
\method{filter}{stars}(.data, ..., .dots)

\method{mutate}{stars}(.data, ..., .dots)

\method{select}{stars}(.data, ...)

\method{pull}{stars}(.data, var = -1)

\method{as.tbl_cube}{stars}(x, ...)
}
\arguments{
\item{.data}{see \link[dplyr]{filter}}

\item{...}{see \link[dplyr]{filter}}

\item{.dots}{see \link[dplyr]{filter}}

\item{var}{see \link[dplyr]{pull}}

\item{x}{object of class \code{stars}}
}
\description{
dplyr verbs for stars objects
}
