% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gdal_utils}
\alias{gdal_utils}
\alias{st_gdalinfo}
\alias{gdal_utils}
\alias{st_gdalwarp}
\alias{gdal_utils}
\alias{st_gdalrasterize}
\title{Native interface to gdal utils}
\usage{
st_gdalinfo(name, options, quiet = FALSE)

st_gdalwarp(source, destination, options)

st_gdalrasterize(source, destination, options)
}
\arguments{
\item{name}{character; name of input layer}

\item{options}{layer opening options}

\item{quiet}{logical; suppress printed output}

\item{source}{character; name(s) of input layer(s)}

\item{destination}{character; name of output layer}
}
\value{
\code{st_gdalwarp} and \code{st_gdalrasterize} return \code{NULL} (invisibly) on success, an error message on failure.
\code{st_gdalinfo} returns the raster metadata
}
\description{
Native interface to gdal utils
}
