% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stars}
\alias{plot.stars}
\alias{plot.stars}
\alias{image.stars}
\title{plot stars object, with subplots for each level of first non-spatial dimension}
\usage{
\method{plot}{stars}(x, y, ..., join_zlim = TRUE, main = names(x)[1],
  axes = FALSE, downsample = TRUE, nbreaks = 11, breaks = "quantile",
  col = grey(1:(nbreaks - 1)/nbreaks))

\method{image}{stars}(x, ..., band = 1, attr = 1, asp = 1, rgb = NULL,
  maxColorValue = max(x[[attr]]), xlab = if (!axes) "" else names(dims)[1],
  ylab = if (!axes) "" else names(dims)[2], xlim = st_bbox(x)$xlim,
  ylim = st_bbox(x)$ylim, text_values = FALSE, axes = FALSE,
  interpolate = FALSE)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{y}{ignored}

\item{...}{for \code{plot}, passed on to \code{image.stars}; for \code{image}, passed on to \code{image.default} or \code{rasterImage}.}

\item{join_zlim}{logical; if \code{TRUE}, compute a single zlim for all subplots from array range}

\item{main}{character; subplot title prefix; use \code{""} to get only time, use \code{NULL} to suppress subplot titles}

\item{axes}{logical; should axes and box be added to the plot?}

\item{downsample}{logical; if \code{TRUE} will try to plot not many more pixels than actually are visibule.}

\item{nbreaks}{number of color breaks; should be one more than number of colors. If missing and \code{col} is specified, it is derived from that.}

\item{breaks}{actual color breaks, or a method name used for \link[classInt]{classIntervals}.}

\item{col}{colors to use for grid cells}

\item{band}{integer; which band (dimension) to plot}

\item{attr}{integer; which attribute to plot}

\item{asp}{numeric; aspect ratio of image}

\item{rgb}{integer; specify three bands to form an rgb composite}

\item{maxColorValue}{numeric; passed on to \link{rgb}}

\item{xlab}{character; x axis label}

\item{ylab}{character; y axis label}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}

\item{text_values}{logical; print values as text on image?}

\item{interpolate}{logical; when using \link{rasterImage} (rgb), should pixels be interpolated?}
}
\description{
plot stars object, with subplots for each level of first non-spatial dimension
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = read_stars(tif)
image(x, col = grey((3:9)/10))
image(x, rgb = c(1,3,5)) # rgb composite
}
