% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R, R/dimensions.R
\name{print_stars}
\alias{print_stars}
\alias{print.stars}
\alias{as.data.frame.dimensions}
\alias{print.dimensions}
\title{print stars or dimensions object}
\usage{
\method{print}{stars}(x, ..., n = 1e+05, abbrev = 30)

\method{as.data.frame}{dimensions}(
  x,
  ...,
  digits = 6,
  usetz = TRUE,
  stars_crs = getOption("stars.crs") \%||\% 28,
  all = FALSE
)

\method{print}{dimensions}(x, ...)
}
\arguments{
\item{x}{object of class stars or of class dimensions}

\item{...}{passed on to \code{as.data.frame.dimensions}}

\item{n}{max number of cells to sample for attribute summary statistics}

\item{abbrev}{number of characters to abbreviate attribute names to}

\item{digits}{number of digits to print numbers}

\item{usetz}{logical; used to format \code{PCICt} or \code{POSIXct} values}

\item{stars_crs}{maximum width of string for CRS objects}

\item{all}{logical; if \code{TRUE} print also fields entirely filled with \code{NA} or \code{NULL}}
}
\description{
print stars or dimensions object
}
