% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{read_stars}
\alias{read_stars}
\alias{read_stars}
\alias{c.stars}
\title{read raster/array dataset from file or connection}
\usage{
read_stars(.x, ..., options = character(0), driver = character(0),
  sub = TRUE, quiet = FALSE, NA_value = NA_real_, along = NA_integer_)

\method{c}{stars}(..., along = NA_integer_)
}
\arguments{
\item{.x}{character vector with name(s) of file(s) or data source(s) to be read}

\item{...}{ignored}

\item{options}{character; opening options}

\item{driver}{character; driver to use for opening file}

\item{sub}{integer or logical; sub-datasets to be read}

\item{quiet}{logical; print progress output?}

\item{NA_value}{numeric value to be used for conversion into NA values; by default this is read from the input file}

\item{along}{length-one character or integer, or list; determines how several arrays are combined, see Details.}
}
\value{
object of class \code{stars}
}
\description{
read raster/array dataset from file or connection
}
\details{
In case \code{.x} contains multiple files, they will all be read and combined with \code{c}. Along which dimension, or how should objects be merged? If \code{along} is set to \code{NA} it will merge arrays as new attributes if all objects have identical dimensions, or else try to merge along time if a dimension called \code{time} indicates different time stamps. A single name (or positive value) for \code{along} will merge along that dimension, or create a new one if it does not already exist. If the arrays should be arranged along one of more dimensions with values (e.g. time stamps), a named list can passed to \code{along} to specify them; see example.
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
(x1 = read_stars(tif))
(x2 = read_stars(c(tif, tif)))
(x3 = read_stars(c(tif, tif), along = "band"))
(x4 = read_stars(c(tif, tif), along = "new_dimensions")) # create 4-dimensional array
x1o = read_stars(tif, options = "OVERVIEW_LEVEL=1")
t1 = as.Date("2018-07-31")
# along is a named list indicating two dimensions:
read_stars(c(tif, tif, tif, tif), along = list(foo = c("bar1", "bar2"), time = c(t1, t1+2)))
}
