% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{read_stars_meta}
\alias{read_stars_meta}
\title{read raster/array metadata from file or connection}
\usage{
read_stars_meta(.x, ..., options = character(0), driver = character(0),
  sub = TRUE, quiet = FALSE, NA_value = NA_real_)
}
\arguments{
\item{.x}{if character, name of file(s) to read; if list: list with arrays}

\item{...}{ignored}

\item{options}{character; opening options}

\item{driver}{character; driver to use for opening file}

\item{sub}{integer or logical; sub-datasets to be read}

\item{quiet}{logical; print progress output?}

\item{NA_value}{numeric value to be used for conversion into NA values; by default this is read from the input file}
}
\value{
object of class \code{stars}
}
\description{
read raster/array dataset from file or connection
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = read_stars(tif)
# x1 = read_stars(nv, options = "OVERVIEW_LEVEL=1")
}
