% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R, R/stars.R
\name{redimension}
\alias{redimension}
\alias{st_redimension.stars_proxy}
\alias{st_redimension}
\alias{st_redimension.stars}
\title{redimension array, or collapse attributes into a new dimension}
\usage{
\method{st_redimension}{stars_proxy}(
  x,
  new_dims = st_dimensions(x),
  along = list(new_dim = names(x)),
  ...
)

st_redimension(x, new_dims, along, ...)

\method{st_redimension}{stars}(
  x,
  new_dims = st_dimensions(x),
  along = setNames(list(names(x)), name),
  ...,
  name = "new_dim"
)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{new_dims}{target dimensions: either a `dimensions` object or an integer vector with the dimensions' sizes}

\item{along}{named list with new dimension name and values}

\item{...}{ignored}

\item{name}{character name of the new dimension}
}
\description{
redimension array, or collapse attributes into a new dimension
}
