% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R, R/raster.R
\name{st_as_stars}
\alias{st_as_stars}
\alias{st_as_stars}
\alias{st_as_stars.Raster}
\title{convert objects into a stars object}
\usage{
st_as_stars(x, ...)

\method{st_as_stars}{Raster}(x, ...)
}
\arguments{
\item{x}{object to convert}

\item{...}{ignored}
}
\description{
convert objects into a stars object
}
