% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R, R/sf.R, R/raster.R
\name{st_as_stars}
\alias{st_as_stars}
\alias{st_as_stars.list}
\alias{st_as_stars.default}
\alias{st_as_stars.sf}
\alias{st_as_stars.Raster}
\title{convert objects into a stars object}
\usage{
st_as_stars(.x, ...)

\method{st_as_stars}{list}(.x, ..., dimensions = NULL)

\method{st_as_stars}{default}(.x = NULL, ...)

\method{st_as_stars}{sf}(.x, ...)

\method{st_as_stars}{Raster}(.x, ...)
}
\arguments{
\item{.x}{object to convert}

\item{...}{ignored}

\item{dimensions}{object of class dimensions}
}
\description{
convert objects into a stars object
}
