% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_crop}
\alias{st_crop}
\title{crop a stars object}
\usage{
st_crop(x, obj, crop = TRUE)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{obj}{object of class \code{sf}, \code{sfc} or \code{bbox}}

\item{crop}{logical; if \code{TRUE}, the spatial extent of the returned object is cropped to still cover \code{obj}}
}
\description{
crop a stars object
}
