% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_dimensions}
\alias{st_dimensions}
\alias{st_dimensions}
\alias{st_dimensions.stars}
\title{get dimensions from stars object}
\usage{
st_dimensions(x, ...)

\method{st_dimensions}{stars}(x, ...)
}
\arguments{
\item{x}{object to retrieve dimensions information from}

\item{...}{further arguments}
}
\value{
the \code{dimensions} attribute of \code{x}, of class \code{dimensions}
}
\description{
get dimensions from stars object
}
