% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_dimensions}
\alias{st_dimensions}
\alias{st_dimensions}
\alias{st_dimensions.stars}
\title{get dimensions from stars object}
\usage{
st_dimensions(x, ...)

\method{st_dimensions}{stars}(x, ...)
}
\description{
get dimensions from stars object
}
