% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{st_geotransform}
\alias{st_geotransform}
\alias{st_geotransform<-}
\title{get or set the geotransform, or rotation matrix}
\usage{
st_geotransform(x, ...)

st_geotransform(x, ...) <- value
}
\arguments{
\item{x}{object of class stars or dimensions}

\item{...}{ignored}

\item{value}{length 6 numeric vector, or 2 x 2 (scaled) rotation matrix}
}
\description{
get or set the geotransform, or rotation matrix
}
