% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_raster_type}
\alias{st_raster_type}
\title{get the raster type (if any) of a stars object}
\usage{
st_raster_type(x)
}
\arguments{
\item{x}{object of class \code{stars}}
}
\value{
one of \code{NA} (if the object does not have raster dimensions), 
\code{"curvilinear"}, \code{"rectilinear"}, \code{"affine"}, or \code{"regular"}
}
\description{
get the raster type (if any) of a stars object
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = read_stars(tif)
st_raster_type(x)
}
