% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_stars}
\alias{st_stars}
\title{read raster/array dataset from file or connection
read raster/array dataset from file or connection}
\usage{
st_stars(file, ...)
}
\arguments{
\item{file}{character; file name to read}

\item{...}{ignored}
}
\value{
object of class \code{stars}
}
\description{
read raster/array dataset from file or connection
read raster/array dataset from file or connection
}
