% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_stars}
\alias{st_stars}
\alias{st_stars}
\alias{st_stars.character}
\alias{st_stars}
\alias{st_stars.list}
\alias{st_stars}
\alias{image.stars}
\title{read raster/array dataset from file or connection}
\usage{
st_stars(x, ...)

\method{st_stars}{character}(x, ..., options = character(0),
  driver = character(0), sub = TRUE, quiet = FALSE)

\method{st_stars}{list}(x, ..., dimensions = NULL)

\method{image}{stars}(x, ..., band = 1, attr = 1, asp = 1, rgb = NULL,
  maxColorValue = 1, xlab = names(dims)[1], ylab = names(dims)[2])
}
\arguments{
\item{x}{if character, file name to read; if list: list with arrays}

\item{...}{passed on to \code{image.default}}

\item{options}{character; opening options}

\item{driver}{character; driver to use for opening file}

\item{sub}{integer or logical; sub-datasets to be read}

\item{quiet}{logical; print progress output?}

\item{dimensions}{object of class dimensions}

\item{band}{integer; which band (dimension) to plot}

\item{attr}{integer; which attribute to plot}

\item{asp}{numeric; aspect ratio of image}

\item{rgb}{integer; specify three bands to form an rgb composite}

\item{maxColorValue}{numeric; passed on to \link{rgb}}

\item{xlab}{character; x axis label}

\item{ylab}{character; y axis label}
}
\value{
object of class \code{stars}
}
\description{
read raster/array dataset from file or connection
}
\examples{
nc = system.file("nc/avhrr-only-v2.19810901.nc", package = "stars")
x = st_stars(nc)
# x1 = st_stars(nv, options = "OVERVIEW_LEVEL=1")
nc = system.file("nc/avhrr-only-v2.19810901.nc", package = "stars")
x = st_stars(nc)
library(abind)
image(adrop(x))
}
