% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R, R/stars.R
\name{st_stars}
\alias{st_stars}
\alias{st_stars.sf}
\alias{st_stars}
\alias{st_stars}
\alias{st_stars.character}
\alias{st_stars}
\alias{st_stars.list}
\title{read raster/array dataset from file or connection}
\usage{
\method{st_stars}{sf}(x, ..., times = colnames(data[[1]]))

st_stars(x, ...)

\method{st_stars}{character}(x, ..., options = character(0),
  driver = character(0), sub = TRUE, quiet = FALSE)

\method{st_stars}{list}(x, ..., dimensions = NULL)
}
\arguments{
\item{x}{if character, file name to read; if list: list with arrays}

\item{...}{ignored}

\item{times}{time instances}

\item{options}{character; opening options}

\item{driver}{character; driver to use for opening file}

\item{sub}{integer or logical; sub-datasets to be read}

\item{quiet}{logical; print progress output?}

\item{dimensions}{object of class dimensions}
}
\value{
object of class \code{stars}
}
\description{
read raster/array dataset from file or connection
}
\examples{
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x = st_stars(tif)
# x1 = st_stars(nv, options = "OVERVIEW_LEVEL=1")
}
