% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdim.R
\name{write_mdim}
\alias{write_mdim}
\title{Write stars object using GDAL multidimensional array interface}
\usage{
write_mdim(
  x,
  filename,
  driver = detect.driver(filename),
  ...,
  root_group_options = character(0),
  options = character(0)
)
}
\arguments{
\item{x}{stars object}

\item{filename}{destination file name}

\item{driver}{character; driver name}

\item{...}{ignored}

\item{root_group_options}{character; driver specific options regarding the creation of the root group}

\item{options}{character; driver specific options regarding the creation of the dataset}
}
\description{
Write stars object using GDAL multidimensional array interface
}
\examples{
set.seed(135)
m = matrix(runif(10), 2, 5)
names(dim(m)) = c("stations", "time")
times = as.Date("2022-05-01") + 1:5
pts = st_as_sfc(c("POINT(0 1)", "POINT(3 5)"))
s = st_as_stars(list(Precipitation = m)) |>
 st_set_dimensions(1, values = pts) |>
 st_set_dimensions(2, values = times)
nc = tempfile(fileext=".nc")
if (compareVersion(sf_extSoftVersion()["GDAL"], "2.5.0") > -1) {
  write_mdim(s, nc)
  # try ncdump on the generated file
  print(read_mdim(nc))
}
}
